+++ system/modules/isotope/languages/en/tl_module.php
@@ -91,11 +91,3 @@ $GLOBALS['TL_LANG']['tl_module']['iso_category_scope_ref']['product']					= 'Cur
 $GLOBALS['TL_LANG']['tl_module']['sortingDirection']['ASC'] 	= "ASC";
 $GLOBALS['TL_LANG']['tl_module']['sortingDirection']['DESC'] 	= "DESC";
 
-
-/**
- * Legends
- */
-$GLOBALS['TL_LANG']['tl_module']['featured_legend']  = 'Featured product settings';
-$GLOBALS['TL_LANG']['tl_module']['config_legend']    = 'Configuration settings';
-$GLOBALS['TL_LANG']['tl_module']['display_legend']    = 'Display settings';
-