+++ system/modules/isotope/CHANGELOG.md
@@ -8,6 +8,7 @@ Version 1.3.0 stable (2012-??-??)
 - PayPal Standard does no longer override the address in debug mode. It could cause issues with international addresses in the sandbox.
 - Use the BYPASS_TOKEN_CHECK constant for postsale script
 - Write log protocol when deleting guest cards
+- Updated palettes and legend labels in tl_module (#2975)
 
 ### Fixed
 - Do not initialize a cart when running the Contao cron job